requirements.txt
fastapi
uvicorn
slack-bolt
openai
python-dotenv
